\name{assignMedianBmiCpp} 
\alias{assignMedianBmiCpp} 
\title{Assign the median BMI to the children.} 
\usage{assignMedianBmiCpp(...)} 
\description{Assign the median BMI to the children.} 
\details{ 
Assign the median BMI to the children. 
} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\value{`vector` of transformed values.} 
\seealso{See also \code{\link{findAgeIndicesCpp}}} 
