\name{calculateCentilesCpp} 
\alias{calculateCentilesCpp} 
\title{Calculation of centiles.} 
\usage{calculateCentilesCpp(...)} 
\description{Calculates the centiles for a vector of z-values.} 
\details{ 
This function converts a vector of z-values to centiles based on the LMS method of Cole (1991). 
} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\value{`vector` of transformed values.} 
\seealso{See also \code{\link{calculateCentileCpp}}, \code{\link{calculateZScoreCpp}}} 
