\name{capwords} 
\alias{capwords} 
\title{Capitalize the first letter of each word} 
\usage{capwords(s, strict)} 
\description{Capitalize the first letter of each word} 
\arguments{ 
\item{s}{A character vector} 
\item{strict}{A logical value. Should only the first letter be capitalized? Default: FALSE. } 
} 
\details{ 
The function is taken from the documentation of the \code{base::chartr} function 
and capitalizes the first letter of each word. 
} 
\value{A character vector} 
