\name{check_columns} 
\alias{check_columns} 
\title{Check for mandatory columns of the input CrescNet data object.} 
\description{ 
Input checking looking for mandatory variable names.  
} 
\details{ 
A check function to look for mandatory variables names in the data object used as an input. 
} 
\arguments{ 
  cresc_data: Input data.frame 
} 
