\name{findIndicesCpp} 
\alias{findIndicesCpp} 
\title{Find the indices for the age values in the reference table.} 
\usage{findIndicesCpp(...)} 
\description{Find the indices for the age values in the reference table.} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\details{ 
Find the indices for the age values in the reference table. 
} 
\value{ 
A `vector` of indices. 
} 
\seealso{See also \code{\link{getIndicesCpp}}.} 
