\name{get_unique_ids_cpp} 
\alias{get_unique_ids_cpp} 
\title{Retrieve unique IDs.} 
\usage{get_unique_ids_cpp(...)} 
\description{Return the unique IDs for a all observations.} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\details{ 
All unique subject IDs are extracted. 
} 
\value{ 
A `vector` of unique subject names. 
} 
\seealso{See also \code{\link{get_indices_cpp}}, \code{\link{get_proportion_cpp}}, \code{\link{lint_cpp}}.} 
