\name{getUniqueIdsCpp} 
\alias{getUniqueIdsCpp} 
\title{Retrieve unique IDs.} 
\usage{getUniqueIdsCpp(...)} 
\description{Return the unique IDs for a all observations.} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\details{ 
All unique subject IDs are extracted. 
} 
\value{ 
A `vector` of unique subject names. 
} 
\seealso{See also \code{\link{getIndicesCpp}}, \code{\link{getProportionCpp}}, \code{\link{interpolateCpp}}.} 
