\name{interpolateCpp} 
\alias{interpolateCpp} 
\title{Linear interpolation between an upper and lower value.} 
\usage{interpolateCpp(...)} 
\description{Perform linear interpolation based on the lower / upper 
reference age bound.} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\details{ 
The basic functionality to perform linear interpolation using the proportion 
of the upper and lower age bound, 
relevant for the respective age of a child. 
} 
\value{A 'numeric' value, representing the linear interpolation.} 
\seealso{See also \code{\link{getIndicesCpp}}, \code{\link{getProportionCpp}}, 
\code{\link{interpolateToZscoreCpp}}} 
