\name{lint_cpp} 
\alias{lint_cpp} 
\title{Linear interpolation between an upper and lower value.} 
\usage{lint_cpp(...)} 
\description{Perform linear interpolation based on the lower / upper reference age bound.} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\details{ 
The basic functionality to perform linear interpolation using the proportion of the upper and lower age bound,  
relevant for the respective age of a child. 
} 
\value{A 'numeric' value, representing the linear interpolation.} 
\seealso{See also \code{\link{get_indices_cpp}}, \code{\link{get_proportion_cpp}}, \code{\link{interpolate_to_z_score_cpp}}} 
