\name{load_ref} 
\alias{load_ref} 
\title{Load reference tables.} 
\description{Using reference tables} 
\arguments{ 
 \item{ref_name}{Either "WHO" or "KiGGS". Default: "WHO".} 
} 
\details{Currently two reference tables are available: 
The World Health Organization (WHO) and the KiGGS based on the study of the Robert-Koch Institute in Germany.} 
\value{A reference as a `data.table` object.} 
