\name{theme_elegant} 
\alias{theme_elegant} 
\title{Custom ggplot theme} 
\description{A customized ggplot theme} 
\usage{ 
theme_elegant(base_size, base_family, axis.line, axis.line.x, axis.line.y, 
legend.title, legend.position, legend.direction, legend.justification, 
legend.box, panel.background, panel.border, panel.grid, plot.background, strip.background, 
strip.text, strip.text.y, ...) 
} 
\arguments{ 
  \item{base_size}{The base font size.} 
  \item{base_family}{The base font family.} 
  \item{axis.line}{The line of the axis.} 
  \item{axis.line.x}{The line of the axis in x direction.} 
  \item{axis.line.y}{The line of the axis in y direction.} 
  \item{legend.title}{The title of the legend.} 
  \item{legend.position}{The position of the legend.} 
  \item{legend.direction}{The direction of the legend.} 
  \item{legend.justification}{The justification of the legend.} 
  \item{legend.box}{The box of the legend.} 
  \item{panel.background}{The background of the panel.} 
  \item{panel.border}{The border of the panel.} 
  \item{panel.grid}{The grid of the panel.} 
  \item{plot.background}{The background of the plot.} 
  \item{strip.background}{The background of the strip.} 
  \item{strip.text}{The text of the strip.} 
  \item{strip.text.y}{The text of the strip in y direction.} 
  \item{...}{Other arguments passed to `theme`.} 
} 
\details{ 
A customized ggplot theme. 
} 
\value{A `ggplot` theme.} 
