\name{theme_elegant_grey} 
\alias{theme_elegant_grey} 
\title{Custom ggplot theme} 
\description{A customized ggplot theme} 
\usage{ 
theme_elegant_grey(base_size, base_family, legend.position, 
legend.title, legend.justification, legend.direction, legend.box, ...) 
} 
\arguments{ 
\item{base_size}{The base font size.} 
\item{base_family}{The base font family.} 
\item{legend.position}{The position of the legend.} 
\item{legend.title}{The title of the legend.} 
\item{legend.justification}{The justification of the legend.} 
\item{legend.direction}{The direction of the legend.} 
\item{legend.box}{The box of the legend.} 
\item{...}{Other arguments passed to `theme`.} 
} 
\details{ 
A customized ggplot theme. 
} 
\value{A `ggplot` theme.} 
