\name{calculateCentilesCpp} 
\alias{calculateCentilesCpp} 
\title{Calculate the centiles for the LMS parameters.} 
\usage{calculateCentilesCpp(...)} 
\description{Calculate the centiles for the LMS parameters.} 
\details{ 
Calculate the centiles for the LMS parameters. 
} 
\arguments{ 
 \item{...}{Arguments.} 
} 
\value{`vector` of transformed values.} 
\seealso{See also \code{\link{calculateZScoreCpp}}} 
