\name{theme_elegant} 
\alias{theme_elegant} 
\title{Custom ggplot theme} 
\description{A customized ggplot theme} 
\usage{theme_elegant(base_size, base_family, legend.title, legend.position)} 
\arguments{ 
  \item{base_size}{The base font size} 
  \item{base_family}{The base font family} 
  \item{legend.title}{The title of the legend} 
  \item{legend.position}{The position of the legend} 
} 
\details{ 
A customized ggplot theme. 
} 
\value{A `ggplot` theme.} 
# A more simplistic, but yet elegant theme for the ggplot 
