\name{theme_elegant} 
\alias{theme_elegant} 
\title{Custom ggplot theme} 
\description{A customized ggplot theme} 
\usage{ 
 theme_elegant(base_size, base_family, legend.title, legend.position,plot.background, strip.text, strip.text.y) 
} 
\arguments{ 
  \item{base_size}{The base font size.} 
  \item{base_family}{The base font family.} 
  \item{legend.title}{The title of the legend.} 
  \item{legend.position}{The position of the legend.} 
  \item{plot.background}{The background of the plot.} 
  \item{strip.text}{The text of the strip.} 
  \item{strip.text.y}{The text of the strip in y direction.} 
} 
\details{ 
A customized ggplot theme. 
} 
\value{A `ggplot` theme.} 
